Extract Downloader into its own file
